// lib/config/interfaces/environment-config.d.ts
export interface SecurityConfig {
    enableWaf: boolean;
    enableGuardDuty: boolean;
    enableConfig: boolean;
    enableCloudTrail: boolean;
    kmsKeyRotation: boolean;
    encryptionAtRest: boolean;
    encryptionInTransit: boolean;
}